ingesta/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.*
google-cloud-bigquery-storage==2.*
pandas==2.*
pyarrow==15.*
requests==2.*
db-dtypes==1.*